requirements.txt
https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.1.73+cuda11.cudnn82-cp38-none-manylinux2010_x86_64.whl
dm-haiku==0.0.5
distrax==0.1.0
rlax==0.0.4
optax==0.0.9
chex==0.0.8
dm-tree
atari-py==0.2.6
gym==0.15.7
ray[tune]==1.8.0
opencv-python
python-snappy